+++ 42_Projects/42_Formation/libft/ft_strmapi.c
@@ -6,7 +6,7 @@
 /*   By: andrefil <andrefil@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/10/18 15:58:13 by andrefil          #+#    #+#             */
-/*   Updated: 2023/10/26 00:30:39 by andrefil         ###   ########.fr       */
+/*   Updated: 2023/11/03 21:51:15 by andrefil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,7 +21,7 @@ char	*ft_strmapi(char const *s, char (*f)(unsigned int, char))
 	if (!result)
 		return (NULL);
 	index = 0;
-	while (index < ft_strlen(s))
+	while (index < (int)ft_strlen(s))
 	{
 		result[index] = f(index, s[index]);
 		index++;
